Refactor SecuredFieldsPure page, split into smaller files and solved linting issues.
