+++ playground/pages/SecuredFieldsPure/SecuredFieldsPure.html
@@ -5,192 +5,14 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <meta http-equiv="X-UA-Compatible" content="ie=edge" />
         <title>Adyen Web | Secured Fields Pure</title>
-        <style>
-            h4 {
-                margin-bottom: 5px;
-            }
-            .caveat {
-                margin: 5px auto;
-                width: 470px;
-                font-family: 'Open Sans', sans-serif;
-                color: #0abf53;
-            }
-            .container {
-                width: 460px;
-                margin: 30px auto;
-            }
-            .payment-div {
-                position: relative;
-                font-family: 'Open Sans', sans-serif;
-                font-size: 14px;
-                height: 220px;
-            }
-            .payment-div-extra {
-                border-top: #999 1px solid;
-            }
-            .payment-div-extra .pm-form-label {
-                margin-top: 10px;
-            }
-            .pm-details {
-                display: block;
-                box-sizing: border-box;
-                padding: 0px 20px 20px 20px;
-                width: 100%;
-            }
-            .pm-form {
-                position: relative;
-                width: 100%;
-                max-width: 420px;
-            }
-            .pm-image {
-                background-color: #ffffff;
-                border-radius: 4px;
-                -moz-boder-radius: 4px;
-                -webkit-border-radius: 4px;
-                float: right;
-                line-height: 0;
-                position: relative;
-                overflow: hidden;
-            }
-            .pm-form-label {
-                float: left;
-                padding-bottom: 1em;
-                position: relative;
-                width: 100%;
-            }
-            .pm-form-label--exp-date {
-                width: 40%;
-            }
-            .pm-form-label--cvc {
-                float: right;
-                width: 40%;
-            }
-            .pm-form-label__text {
-                color: #00112c;
-                float: left;
-                font-size: 0.93333em;
-                padding-bottom: 6px;
-                position: relative;
-            }
-            .pm-input-field {
-                background: white;
-                border: 1px solid #d8d8d8;
-                -moz-border-radius: 4px;
-                -webkit-border-radius: 4px;
-                border-radius: 4px;
-                box-sizing: border-box;
-                clear: left;
-                font-size: 0.93333333333em;
-                float: left;
-                padding: 8px;
-                position: relative;
-                width: 100%;
-                height: 35px;
-            }
-            .pm-input-field-name:focus,
-            .pm-input-field-other:focus {
-                border: 1px solid #969696;
-                outline: none;
-            }
-            .pm-input-field--error {
-                border: 1px solid #ff7d00;
-            }
-            .pm-input-field--focus,
-            .sf-ios-additional-el-focus {
-                border: 1px solid #969696;
-                outline: none;
-            }
-            .pm-input-field--error.pm-input-field--focus {
-                border: 1px solid #ff7d00;
-            }
-            .pm-form-label__error-text {
-                color: #ff7d00;
-                display: none;
-                float: left;
-                font-size: 13px;
-                padding-top: 0.4em;
-                position: relative;
-                width: 100%;
-            }
-            .pm-button--submit {
-                background-color: #30c94a;
-                border: 0;
-                width: 420px;
-                margin: 0px 20px 20px;
-                pointer-events: none;
-                opacity: 0.3;
-                -moz-border-radius: 2px;
-                -webkit-border-radius: 2px;
-                border-radius: 2px;
-                color: white;
-                font-size: 1em;
-                font-weight: 500;
-                padding: 20px 0;
-                position: relative;
-            }
-            .pm-button--submit:hover {
-                background-color: #00914b;
-                cursor: auto;
-            }
-            .pm-button--submit:focus {
-                outline: none;
-            }
-            .is-ios * {
-                cursor: pointer;
-            }
-
-            .adyen-checkout__spinner__wrapper {
-                height: 100%;
-                display: flex;
-                justify-content: center;
-                align-items: center;
-            }
-
-            .adyen-checkout__spinner__wrapper--inline {
-                height: auto;
-                display: inline-block;
-                margin-right: 8px;
-            }
-
-            .adyen-checkout__spinner {
-                border: 3px solid #fff;
-                border-radius: 50%;
-                height: 43px;
-                width: 43px;
-                border-top-color: transparent;
-                animation: rotateSpinner 1.5s infinite linear;
-            }
-
-            .adyen-checkout__spinner--large {
-                height: 43px;
-                width: 43px;
-            }
-
-            .adyen-checkout__spinner--small {
-                height: 16px;
-                width: 16px;
-                border-width: 2px;
-            }
-
-            .adyen-checkout__spinner--medium {
-                height: 28px;
-                width: 28px;
-            }
-
-            @keyframes rotateSpinner {
-                0% {
-                    transform: rotate(0deg);
-                }
-                100% {
-                    transform: rotate(360deg);
-                }
-            }
-        </style>
     </head>
-    <body style="max-width: 100%;">
+    <body>
         <header></header>
         <main>
-            <h5 class="caveat">For Internal Dev only (works only if you can host SecuredFields locally)</h5>
+            <div class="info">
+                For Internal Dev only (works only if you can host SecuredFields locally)
+            </div>
+
             <div class="container">
                 <div id="adyenPaymentDivCard1" class="payment-div">
                     <h4>Card</h4>
